requirements.txt
certifi==2025.1.31
influxdb-client==1.48.0
influxdb3-python==0.12.0
networkx==3.4.2
pyarrow==19.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
reactivex==4.0.4
typing-extensions==4.13.2
urllib3==2.4.0
